+++ src/ui/camera.test.ts
@@ -1,6 +1,7 @@
 import {describe, beforeEach, test, expect, vi} from 'vitest';
 import {Camera, type CameraOptions, type PointLike} from '../ui/camera';
 import {TaskQueue, type TaskID} from '../util/task_queue';
+import * as timeControl from '../util/time_control';
 import {browser} from '../util/browser';
 import {fixedLngLat, fixedNum} from '../../test/unit/lib/fixed';
 import {setMatchMedia} from '../util/test/util';
@@ -1042,7 +1043,7 @@ describe('easeTo', () => {
 
     test('can be called from within a moveend event handler', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         stub.mockImplementation(() => 0);
         camera.easeTo({center: [100, 0], duration: 10});
@@ -1079,7 +1080,7 @@ describe('easeTo', () => {
 
     test('pans eastward across the antimeridian', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         camera.setCenter([170, 0]);
         let crossedAntimeridian;
@@ -1122,7 +1123,7 @@ describe('easeTo', () => {
 
     test('pans westward across the antimeridian', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         camera.setCenter([-170, 0]);
         let crossedAntimeridian;
@@ -1166,7 +1167,7 @@ describe('easeTo', () => {
     test('animation occurs when prefers-reduced-motion: reduce is set but overridden by essential: true', async () => {
         const camera = createCamera();
         Object.defineProperty(browser, 'prefersReducedMotion', {value: true});
-        const stubNow = vi.spyOn(browser, 'now');
+        const stubNow = vi.spyOn(timeControl, 'now');
 
         // camera transition expected to take in this range when prefersReducedMotion is set and essential: true,
         // when a duration of 200 is requested
@@ -1174,7 +1175,7 @@ describe('easeTo', () => {
         const max = 300;
 
         let startTime;
-        camera.on('movestart', () => { startTime = browser.now(); });
+        camera.on('movestart', () => { startTime = timeControl.now(); });
         const promise = camera.once('moveend');
 
         setTimeout(() => {
@@ -1190,7 +1191,7 @@ describe('easeTo', () => {
         }, 0);
 
         await promise;
-        const endTime = browser.now();
+        const endTime = timeControl.now();
         const timeDiff = endTime - startTime;
         expect(timeDiff >= min && timeDiff < max).toBeTruthy();
     });
@@ -1281,7 +1282,7 @@ describe('easeTo', () => {
 
     test('terrain set during easeTo', () => {
         const camera = createCamera();
-        const stubNow = vi.spyOn(browser, 'now');
+        const stubNow = vi.spyOn(timeControl, 'now');
 
         stubNow.mockImplementation(() => 0);
 
@@ -1344,7 +1345,7 @@ describe('flyTo', () => {
     test('Zoom out from the same position to the same position with animation', async () => {
         const pos = {lng: 0, lat: 0};
         const camera = createCamera({zoom: 20, center: pos});
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         const promise = camera.once('moveend');
 
@@ -1538,7 +1539,7 @@ describe('flyTo', () => {
         camera.on('pitchend', (d) => { pitchended = d.data; });
         const promise = camera.once('moveend');
 
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
         stub.mockImplementation(() => 0);
 
         camera.flyTo({center: [100, 0], duration: 10}, eventData);
@@ -1579,7 +1580,7 @@ describe('flyTo', () => {
     });
 
     test('no roll when motion is interrupted', () => {
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         const camera = createCamera();
         stub.mockImplementation(() => 0);
@@ -1591,7 +1592,7 @@ describe('flyTo', () => {
     });
 
     test('no roll when motion is interrupted: globe', () => {
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         const camera = createCameraGlobe();
         stub.mockImplementation(() => 0);
@@ -1603,7 +1604,7 @@ describe('flyTo', () => {
     });
 
     test('angles when motion is interrupted', () => {
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         const camera = createCamera();
         stub.mockImplementation(() => 0);
@@ -1617,7 +1618,7 @@ describe('flyTo', () => {
     });
 
     test('angles when motion is interrupted: globe', () => {
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         const camera = createCameraGlobe();
         stub.mockImplementation(() => 0);
@@ -1632,7 +1633,7 @@ describe('flyTo', () => {
 
     test('can be called from within a moveend event handler', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
         stub.mockImplementation(() => 0);
 
         camera.flyTo({center: [100, 0], duration: 10});
@@ -1673,7 +1674,7 @@ describe('flyTo', () => {
 
         const promise = camera.once('moveend');
 
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
         stub.mockImplementation(() => 0);
 
         camera.flyTo({center: [100, 0], zoom: 18, duration: 10});
@@ -1693,7 +1694,7 @@ describe('flyTo', () => {
 
     test('pans eastward across the prime meridian', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         camera.setCenter([-10, 0]);
         let crossedPrimeMeridian;
@@ -1725,7 +1726,7 @@ describe('flyTo', () => {
 
     test('pans westward across the prime meridian', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         camera.setCenter([10, 0]);
         let crossedPrimeMeridian;
@@ -1757,7 +1758,7 @@ describe('flyTo', () => {
 
     test('pans eastward across the antimeridian', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         camera.setCenter([170, 0]);
         let crossedAntimeridian;
@@ -1788,7 +1789,7 @@ describe('flyTo', () => {
 
     test('pans westward across the antimeridian', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         camera.setCenter([-170, 0]);
         let crossedAntimeridian;
@@ -1819,7 +1820,7 @@ describe('flyTo', () => {
 
     test('does not pan eastward across the antimeridian if no world copies', async () => {
         const camera = createCamera({renderWorldCopies: false});
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         camera.setCenter([170, 0]);
         let crossedAntimeridian;
@@ -1851,7 +1852,7 @@ describe('flyTo', () => {
 
     test('does not pan westward across the antimeridian if no world copies', async () => {
         const camera = createCamera({renderWorldCopies: false});
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         camera.setCenter([-170, 0]);
         let crossedAntimeridian;
@@ -1883,7 +1884,7 @@ describe('flyTo', () => {
 
     test('jumps back to world 0 when crossing the antimeridian', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         camera.setCenter([-170, 0]);
 
@@ -1914,7 +1915,7 @@ describe('flyTo', () => {
 
     test('peaks at the specified zoom level', async () => {
         const camera = createCamera({zoom: 20});
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         const minZoom = 1;
         let zoomed = false;
@@ -1958,7 +1959,7 @@ describe('flyTo', () => {
 
         const promise = camera.once('moveend');
 
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
         stub.mockImplementation(() => 0);
         camera.flyTo({center: [12, 34], zoom: 30, duration: 10});
 
@@ -1983,7 +1984,7 @@ describe('flyTo', () => {
 
         const promise = camera.once('moveend');
 
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
         stub.mockImplementation(() => 0);
         camera.flyTo({center: [12, 34], zoom: 1, duration: 10});
 
@@ -2031,7 +2032,7 @@ describe('flyTo', () => {
 
     test('check elevation events freezeElevation=false', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         const terrainCallbacks = {prepare: 0, update: 0, finalize: 0} as any;
         camera.terrain = {} as Terrain;
@@ -2055,7 +2056,7 @@ describe('flyTo', () => {
 
     test('check elevation events freezeElevation=true', async() => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
 
         const terrainCallbacks = {prepare: 0, update: 0, finalize: 0} as any;
         camera.terrain = {} as Terrain;
@@ -2121,7 +2122,7 @@ describe('isEasing', () => {
     test('returns false when done panning', async () => {
         const camera = createCamera();
         const promise = camera.once('moveend');
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
         stub.mockImplementation(() => 0);
         camera.panTo([100, 0], {duration: 1});
         setTimeout(() => {
@@ -2143,7 +2144,7 @@ describe('isEasing', () => {
     test('returns false when done zooming', async () => {
         const camera = createCamera();
         const promise = camera.once('moveend');
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
         stub.mockImplementation(() => 0);
         camera.zoomTo(3.2, {duration: 1});
         setTimeout(() => {
@@ -2164,7 +2165,7 @@ describe('isEasing', () => {
     test('returns false when done rotating', async () => {
         const camera = createCamera();
         const promise = camera.once('moveend');
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
         stub.mockImplementation(() => 0);
         camera.rotateTo(90, {duration: 1});
         setTimeout(() => {
@@ -2239,7 +2240,7 @@ describe('stop', () => {
         const spy = vi.fn();
         camera.on('moveend', spy);
 
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
         stub.mockImplementation(() => 0);
         camera.panTo([100, 0], {duration: 1}, eventData);
 
@@ -2514,7 +2515,7 @@ describe('transformCameraUpdate', () => {
 
     test('invoke transformCameraUpdate callback during easeTo', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
         stub.mockImplementation(() => 0);
 
         let callbackCount = 0;
@@ -2548,7 +2549,7 @@ describe('transformCameraUpdate', () => {
 
     test('invoke transformCameraUpdate callback during flyTo', async () => {
         const camera = createCamera();
-        const stub = vi.spyOn(browser, 'now');
+        const stub = vi.spyOn(timeControl, 'now');
         stub.mockImplementation(() => 0);
 
         let callbackCount = 0;
@@ -2858,7 +2859,7 @@ describe('easeTo globe projection', () => {
 
         test('smoothly sets given padding with duration > 0', async () => {
             const camera = createCameraGlobe();
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
             const promise = camera.once('moveend');
 
             stub.mockImplementation(() => 0);
@@ -3020,7 +3021,7 @@ describe('easeTo globe projection', () => {
 
         test('pans eastward across the antimeridian', async () => {
             const camera = createCameraGlobe();
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
 
             camera.setCenter([170, 0]);
             let crossedAntimeridian;
@@ -3062,7 +3063,7 @@ describe('easeTo globe projection', () => {
 
         test('pans westward across the antimeridian', async () => {
             const camera = createCameraGlobe();
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
 
             camera.setCenter([-170, 0]);
             let crossedAntimeridian;
@@ -3162,7 +3163,7 @@ describe('flyTo globe projection', () => {
         test('Zoom out from the same position to the same position with animation', async () => {
             const pos = {lng: 0, lat: 0};
             const camera = createCameraGlobe({zoom: 20, center: pos});
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
 
             const promise = camera.once('zoomend');
 
@@ -3241,7 +3242,7 @@ describe('flyTo globe projection', () => {
 
         test('smoothly sets given padding with duration > 0', async () => {
             const camera = createCameraGlobe();
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
             const promise = camera.once('moveend');
 
             stub.mockImplementation(() => 0);
@@ -3408,7 +3409,7 @@ describe('flyTo globe projection', () => {
             camera.on('pitchend', (d) => { pitchended = d.data; });
             const promise = camera.once('moveend');
 
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
             stub.mockImplementation(() => 0);
 
             camera.flyTo({center: [100, 0], duration: 10}, eventData);
@@ -3455,7 +3456,7 @@ describe('flyTo globe projection', () => {
 
             const promise = camera.once('moveend');
 
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
             stub.mockImplementation(() => 0);
 
             camera.flyTo({center: [100, 0], zoom: 18, duration: 10});
@@ -3476,7 +3477,7 @@ describe('flyTo globe projection', () => {
 
         test('pans eastward across the prime meridian', async () => {
             const camera = createCameraGlobe();
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
 
             camera.setCenter([-10, 0]);
             let crossedPrimeMeridian;
@@ -3508,7 +3509,7 @@ describe('flyTo globe projection', () => {
 
         test('pans westward across the prime meridian', async () => {
             const camera = createCameraGlobe();
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
 
             camera.setCenter([10, 0]);
             let crossedPrimeMeridian;
@@ -3540,7 +3541,7 @@ describe('flyTo globe projection', () => {
 
         test('pans eastward across the antimeridian', async () => {
             const camera = createCameraGlobe();
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
 
             camera.setCenter([170, 0]);
             let crossedAntimeridian;
@@ -3572,7 +3573,7 @@ describe('flyTo globe projection', () => {
 
         test('pans westward across the antimeridian', async () => {
             const camera = createCameraGlobe();
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
 
             camera.setCenter([-170, 0]);
             let crossedAntimeridian;
@@ -3604,7 +3605,7 @@ describe('flyTo globe projection', () => {
 
         test('pans eastward across the antimeridian even if renderWorldCopies: false', async () => {
             const camera = createCameraGlobe({renderWorldCopies: false});
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
 
             camera.setCenter([170, 0]);
             let crossedAntimeridian;
@@ -3636,7 +3637,7 @@ describe('flyTo globe projection', () => {
 
         test('pans westward across the antimeridian even if renderWorldCopies: false', async () => {
             const camera = createCameraGlobe({renderWorldCopies: false});
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
 
             camera.setCenter([-170, 0]);
             let crossedAntimeridian;
@@ -3668,7 +3669,7 @@ describe('flyTo globe projection', () => {
 
         test('jumps back to world 0 when crossing the antimeridian', async () => {
             const camera = createCameraGlobe();
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
 
             camera.setCenter([-170, 0]);
 
@@ -3699,7 +3700,7 @@ describe('flyTo globe projection', () => {
 
         test('peaks at the specified zoom level', async () => {
             const camera = createCameraGlobe({zoom: 20});
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
 
             const minZoom = 1;
             let zoomed = false;
@@ -3743,7 +3744,7 @@ describe('flyTo globe projection', () => {
 
             const promise = camera.once('moveend');
 
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
             stub.mockImplementation(() => 0);
             camera.flyTo({center: [12, 34], zoom: 30, duration: 10});
 
@@ -3772,7 +3773,7 @@ describe('flyTo globe projection', () => {
 
             const promise = camera.once('moveend');
 
-            const stub = vi.spyOn(browser, 'now');
+            const stub = vi.spyOn(timeControl, 'now');
             stub.mockImplementation(() => 0);
             camera.flyTo({center: target, zoom: 1, duration: 10});